+++ client-app/src/admin/tests/index.ts
@@ -8,5 +8,6 @@ export * from './gridScrolling/GridScrolling';
 export * from './localDate/LocalDateTestPanel';
 export * from './panels/PanelResizingTestPanel';
 export * from './select/SelectTestPanel';
+export * from './storeProxy/StoreProxyBenchPanel';
 export * from './viewmanager/ViewManagerTestPanel';
 export * from './websocket/WebSocketTestPanel';